arceo/pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "arceo"
version = "0.1.0"
description = "Monitor any AI agent's tool calls and get a risk report."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = []

[project.optional-dependencies]
langchain = ["langchain-core>=0.2"]
crewai = ["crewai>=0.50"]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.30"]
backend = ["httpx>=0.27"]
cli = ["click>=8.0", "pyyaml>=6.0", "httpx>=0.27"]
all = [
    "langchain-core>=0.2",
    "crewai>=0.50",
    "openai>=1.0",
    "anthropic>=0.30",
    "httpx>=0.27",
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.scripts]
arceo = "arceo.cli:main"

[tool.setuptools.packages.find]
include = ["arceo*"]
